pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.71.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pymupdf>=1.26.5",
    "python-dotenv>=1.1.1",
]
